feat(diff): adds field to select heads-only as DiffIncluder option
Currently, the default functionaliy when no versions or bundles are
specified for individual packges, the whole package is added to the
output. The same goes for channels. While preserving this as a default,
the HeadsOnly field has been added to the DiffInclude to allow channel
heads for packages to be selected as an option.

Signed-off-by: Jennifer Power <barnabei.jennifer@gmail.com>